// DXC_Frontend/src/api/models/tinTucArticleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface TinTucArticleDto {
  id?: number
  publicId?: string
  /** @nullable */
  title: string | null
  /** @nullable */
  summary?: string | null
  /** @nullable */
  thumbnailUrl?: string | null
  categoryId?: number
  /** @nullable */
  categoryName?: string | null
  currentStatusId?: number
  /** @nullable */
  currentStatusCode?: string | null
  /** @nullable */
  currentStatusName?: string | null
  /** @nullable */
  currentStatusColor?: string | null
  /** @nullable */
  authorUserPublicId?: string | null
  /** @nullable */
  authorName?: string | null
  viewCount?: number
  isPublic?: boolean
  /** @nullable */
  publishedAt?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}